README.md
This project is a step forward in intelligent healthcare — designed to group patients based on symptoms using Unsupervised Learning (Clustering). The goal is to identify patterns in symptom data that can help in the early detection and categorization of diseases.Helpful in hospitals, clinics, or health apps to pre-analyze patient symptom data and understand disease grouping without relying on labeled data.